% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlib_auto_delin.R
\name{dlib_auto_delin}
\alias{dlib_auto_delin}
\title{dlib Auto-Delineation}
\usage{
dlib_auto_delin(
  stimuli,
  model = c("dlib7", "dlib70"),
  replace = FALSE,
  dlib_path = NULL
)
}
\arguments{
\item{stimuli}{list of class stimlist}

\item{model}{Which of the built-in models (dlib7, dlib70)}

\item{replace}{if FALSE, only gets templates for images with no template}

\item{dlib_path}{path to a custom dlib .dat landmark file to use (model is ignored if set)}
}
\value{
stimlist with templates
}
\description{
Automatically delineate faces using dlib shape predictor models.
}
\examples{
\dontrun{
  auto_dlib7 <- demo_stim() |>
    auto_delin(replace = TRUE) # replace existing templates

  auto_dlib70 <- demo_stim() |>
    auto_delin(model = "dlib70", replace = TRUE)
}
}
