% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_to_tem.R
\name{xml_to_tem}
\alias{xml_to_tem}
\title{XML to Templates}
\usage{
xml_to_tem(xml)
}
\arguments{
\item{xml}{The file path for the XML file}
}
\value{
A stimlist with only tem points
}
\description{
Make webmorph templates out of XML formatted point files.
}
\examples{
\dontrun{
  # requires python and dlib
  xml <- system.file("demo/_images.xml", package = "webmorphR.dlib")
  stimuli <- xml_to_tem(xml)
  stimuli |> draw_tem() |> plot(nrow = 2)
}
}
